moved connection files to middleware and remapped imports
