+++ src/api/v1/requests/index.js
@@ -1,4 +1,4 @@
-import constants from './constants';
+import constants from '../constants';
 
 export default {
     transactions: {
