+++ src/api/v1/resolvers/balanceCurrent.js
@@ -1,4 +1,4 @@
-import statusResolver from 'src/api/v1/statusResolver';
+import statusResolver from 'src/middleware/statusResolver';
 
 const balanceCurrent = ({ data, status }) => {
     if (status !== 200) {
