+++ src/api/v1/resolvers/spentToday.js
@@ -1,4 +1,4 @@
-import statusResolver from 'src/api/v1/statusResolver';
+import statusResolver from 'src/middleware/statusResolver';
 
 const spentToday = ({ data, status }) => {
     if (status !== 200) {
